Fix display of high-precision Monetary (#80)

+++ src/CurrencyFormatting/render.jl
@@ -85,8 +85,8 @@ function render(template::Vector, spec::FormatSpecification, m::Monetary)
     digisep = get(spec, DigitSeparator, DigitSeparator(""))
 
     prec = decimals(m)
-    intpart = trunc(Int, abs(m.val))
-    floatpart = round(Int, big(10)^prec * (abs(m.val) - intpart))
+    intpart = trunc(BigInt, abs(m.val))
+    floatpart = round(BigInt, big(10)^prec * (abs(m.val) - intpart))
 
     syms = getsymboltable(spec)
     next_template = []
